refactor(design-system): internalize cn into the DS boundary (roadmap 1.2)
Moves cn from the app-root lib/utils.ts into nextjs-app/shared/lib/cn.ts so it
travels with the catalog when it becomes a package. Repoints all 55 component-
tree imports to a relative path (survives the later workspace move); lib/utils.ts
re-exports cn so app-level `@/lib/utils` call sites are unchanged. Single source,
no duplication.

Catalog @/ coupling: catalogAppImports 41 -> 5 (ratchet tightened to 5). The
remaining 5 are AnimationProvider/gsap/pseo, handled in later Phase 1 tasks.

Local gate green: typecheck, lint, lint:css, test 2084/2084, build.
check:astryx-roadmap green.
