+++ nextjs-app/shared/components/ArticleShareSection/ArticleShareSection.tsx
@@ -9,7 +9,7 @@ import {
   Link,
   Check,
 } from "@phosphor-icons/react";
-import { cn } from "@/lib/utils";
+import { cn } from "../../lib/cn";
 
 export interface ArticleShareSectionProps {
   /** Article URL to share */
