+++ nextjs-app/shared/components/AspectRatio/AspectRatio.tsx
@@ -1,5 +1,5 @@
 import { type ReactNode } from "react";
-import { cn } from "@/lib/utils";
+import { cn } from "../../lib/cn";
 
 export interface AspectRatioProps {
   /** Media or placeholder content inside the ratio box. */
