+++ nextjs-app/shared/components/AvatarGroup/AvatarGroup.tsx
@@ -1,5 +1,5 @@
 import React, { Children } from "react";
-import { cn } from "@/lib/utils";
+import { cn } from "../../lib/cn";
 import styles from "./AvatarGroup.module.css";
 
 export type AvatarGroupSize = "sm" | "md" | "lg";
