+++ nextjs-app/shared/components/BlogMediaImage/BlogMediaImage.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import Image from "next/image";
-import { cn } from "@/lib/utils";
+import { cn } from "../../lib/cn";
 import { isSvgSrc } from "@/lib/media/imageSrc";
 
 export interface BlogMediaImageProps {
