+++ nextjs-app/shared/components/ButtonGroup/ButtonGroup.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { cn } from "@/lib/utils";
+import { cn } from "../../lib/cn";
 import styles from "./ButtonGroup.module.css";
 
 export interface ButtonGroupProps extends React.HTMLAttributes<HTMLDivElement> {
