+++ nextjs-app/shared/components/Center/Center.tsx
@@ -1,5 +1,5 @@
 import { type ReactNode, type ElementType } from "react";
-import { cn } from "@/lib/utils";
+import { cn } from "../../lib/cn";
 
 export interface CenterProps {
   /** Content to center inside the flex container. */
