+++ nextjs-app/shared/components/CodeBlockWindow/CodeBlockWindow.tsx
@@ -4,7 +4,7 @@ import React, { useEffect, useMemo, useRef, useState } from "react";
 import { useTranslation } from "react-i18next";
 import Button from "@dt/Button";
 import Text from "@dt/Text";
-import { cn } from "@/lib/utils";
+import { cn } from "../../lib/cn";
 import styles from "./CodeBlockWindow.module.css";
 
 export interface CodeBlockWindowProps {
