+++ nextjs-app/shared/components/Container/Container.tsx
@@ -1,5 +1,5 @@
 import { type ReactNode, type ElementType } from "react";
-import { cn } from "@/lib/utils";
+import { cn } from "../../lib/cn";
 
 import styles from "./Container.module.css";
