+++ nextjs-app/shared/components/DashLeadList/DashLeadList.tsx
@@ -1,6 +1,6 @@
 import type { ReactNode } from "react";
 
-import { cn } from "@/lib/utils";
+import { cn } from "../../lib/cn";
 
 import styles from "./DashLeadList.module.css";
