+++ nextjs-app/shared/components/Display/Display.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { cn } from "@/lib/utils";
+import { cn } from "../../lib/cn";
 
 export interface DisplayProps extends React.HTMLAttributes<HTMLElement> {
   /** Element to render, defaults to h1 */
