+++ nextjs-app/shared/components/FormField/FormField.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { type ReactNode } from "react";
-import { cn } from "@/lib/utils";
+import { cn } from "../../lib/cn";
 import styles from "./FormField.module.css";
 
 export interface FormFieldProps {
