+++ nextjs-app/shared/components/GradientDots/GradientDots.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { cn } from "@/lib/utils";
+import { cn } from "../../lib/cn";
 import styles from "./GradientDots.module.css";
 
 export interface GradientDotsProps {
