+++ nextjs-app/shared/components/LanguageSwitcher/LanguageSwitcher.tsx
@@ -3,7 +3,7 @@
 import { useCallback, useEffect, useId, useRef, useState } from "react";
 import { AnimatePresence, motion, useReducedMotion } from "framer-motion";
 import { useTranslation } from "react-i18next";
-import { cn } from "@/lib/utils";
+import { cn } from "../../lib/cn";
 import styles from "./LanguageSwitcher.module.css";
 
 export type LanguageSwitcherOption = {
