+++ nextjs-app/shared/components/MdxImage/MdxImage.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { useState } from "react";
-import { cn } from "@/lib/utils";
+import { cn } from "../../lib/cn";
 import { isSvgSrc } from "@/lib/media/imageSrc";
 import { BlogMediaImage } from "@dt/BlogMediaImage";
