+++ nextjs-app/shared/components/NavLink/NavLink.tsx
@@ -3,7 +3,7 @@
 import Link from "next/link";
 import { usePathname } from "next/navigation";
 import { type ReactNode } from "react";
-import { cn } from "@/lib/utils";
+import { cn } from "../../lib/cn";
 
 export interface NavLinkProps {
   /** Route path for next/link. */
