+++ nextjs-app/shared/components/Pagination/Pagination.tsx
@@ -3,7 +3,7 @@
 import { useMemo } from "react";
 import { useTranslation } from "react-i18next";
 import { ChevronLeft, ChevronRight } from "lucide-react";
-import { cn } from "@/lib/utils";
+import { cn } from "../../lib/cn";
 
 export interface PaginationProps {
   /** Current active page (1-indexed) */
