+++ nextjs-app/shared/components/Prose/Prose.tsx
@@ -1,5 +1,5 @@
 import { type ReactNode } from "react";
-import { cn } from "@/lib/utils";
+import { cn } from "../../lib/cn";
 
 export interface ProseProps {
   children: ReactNode;
