+++ nextjs-app/shared/components/ScrollIndicator/ScrollIndicator.tsx
@@ -4,7 +4,7 @@ import { useRef, useCallback } from "react";
 import { ChevronDown } from "lucide-react";
 import { gsap, useGSAP } from "@/nextjs-app/shared/lib/gsap";
 import { useAnimationContext } from "@/providers/AnimationProvider";
-import { cn } from "@/lib/utils";
+import { cn } from "../../lib/cn";
 
 export interface ScrollIndicatorProps {
   /** ID of the target element to scroll to */
