+++ nextjs-app/shared/components/SkillsGrid/SkillsGrid.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { type ReactNode, useMemo } from "react";
-import { cn } from "@/lib/utils";
+import { cn } from "../../lib/cn";
 import {
   Tooltip,
   TooltipContent,
