+++ nextjs-app/shared/components/Spacer/Spacer.tsx
@@ -1,4 +1,4 @@
-import { cn } from "@/lib/utils";
+import { cn } from "../../lib/cn";
 
 export interface SpacerProps {
   /** Tokenized gap size. @default "md" */
