+++ nextjs-app/shared/components/Stack/Stack.tsx
@@ -1,5 +1,5 @@
 import { type ReactNode, type ElementType } from "react";
-import { cn } from "@/lib/utils";
+import { cn } from "../../lib/cn";
 
 export interface StackProps {
   /** Stacked children. */
