+++ nextjs-app/shared/components/TableOfContents/TableOfContents.tsx
@@ -3,7 +3,7 @@
 import { useState } from "react";
 import { useTranslation } from "react-i18next";
 import { ChevronDown } from "lucide-react";
-import { cn } from "@/lib/utils";
+import { cn } from "../../lib/cn";
 
 export interface TOCItem {
   /** Heading element ID */
