+++ nextjs-app/shared/components/ValueCard/ValueCard.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { type ReactNode } from "react";
-import { cn } from "@/lib/utils";
+import { cn } from "../../lib/cn";
 
 export interface ValueCardProps {
   /** Icon element */
