+++ nextjs-app/shared/components/animations/AnimatedFormField/AnimatedFormField.tsx
@@ -3,7 +3,7 @@
 import { useRef, useCallback, type ReactNode } from "react";
 import { gsap, useGSAP } from "@/nextjs-app/shared/lib/gsap";
 import { useAnimationContext } from "@/providers/AnimationProvider";
-import { cn } from "@/lib/utils";
+import { cn } from "../../../lib/cn";
 import styles from "./AnimatedFormField.module.css";
 
 export interface AnimatedFormFieldProps {
